Redirection Update
The Redirection Update adds Redirection page that will fetch user data in the database, find if there is some informations, and will redirect the user in one of the different pages.

Multiples security layers will be added in the future to prevent any kind of usurpations, and the API will be securized with tokens to access it.