Register Update V0.21
Register Update V0.21 adds the register page's image, at the bottom left of the screen. It pushes the "Next" button a bit on the right.
Changes also the space between the element, which was necessary to implement the image well.
Changes the Row of the 1st Text "Register" into a Container to margin the text, which was too high up.